asklegal_enhanced/requirements.txt
fastapi>=0.104.0,<0.105.0
uvicorn>=0.24.0,<0.25.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
redis>=5.0.0,<6.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-multipart>=0.0.6,<0.0.7
python-jose>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
python-dotenv>=1.0.0,<2.0.0
PyPDF2>=3.0.1,<4.0.0
python-docx>=1.2.0,<2.0.0
sentence-transformers>=2.2.2,<3.0.0
pillow>=10.0.0,<11.0.0
ctransformers>=0.2.27,<0.3.0
requests>=2.31.0,<3.0.0
numpy>=1.26.2,<2.0.0
pdfplumber>=0.10.0,<0.11.0
pytesseract>=0.3.10,<0.4.0
transformers>=4.30.0,<5.0.0
torch>=2.0.0,<3.0.0
spacy>=3.5.0,<4.0.0
neo4j>=5.0.0,<6.0.0
layoutparser>=0.3.4,<0.4.0
unstructured>=0.10.0,<0.11.0